requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.14.0
idna==3.10
keras==2.15.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pillow==10.3.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.6.0
wrapt==1.14.2
